% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baq_niplot.R
\name{baq_niplot}
\alias{baq_niplot}
\title{Plot News Impact}
\usage{
baq_niplot(baq_nif, ni_plots = c("var1", "var2", "cor"),
  plotdir = "horizontal")
}
\arguments{
\item{baq_nif}{A baq_nif class object.}

\item{ni_plots}{A character vector specifying which conditional correlation
/ variance to plot. Can contain any combination of "var1", "var2" and
"cor". Defaults to c("var1", "var2", "cor").}

\item{plotdir}{A character vector of length one to determine if the plots
should be created in a row ("horizontal") or a column ("vertical").
Defaults to "horizontal".}
}
\value{
A plot of the News Impact on the conditional variance / correlation.
}
\description{
Plot the News Impact on the conditional variance / correlation of a baq_nif
class object.
}
\examples{
# create data
eps <- mgarchBEKK::simulateBEKK(2, 100)

# fit the model
gjr <- mgarchBEKK::mGJR(eps$eps[[1]], eps$eps[[2]])

# apply the news impact function to the model
nif <- baq_nifunction(gjr)

# plot the results
baq_niplot(nif)

# plot only the the first series
baq_niplot(nif, ni_plots = "var1")

\dontrun{

# save plot as pdf
pdf("baqplot_example.pdf", width = 15, height = 7.2)
baq_niplot(nif)
dev.off()
}
}
\references{
{
  H. Schmidbauer & A. Roesch. Volatility Spillovers Between Crude Oil
  Prices. International Conference on Policy Modeling. EcoMod,
  Berlin, 2008.

  H. Schmidbauer & A. Roesch. Volatility Spillovers Between Crude Oil
  Prices and Us Dollar To Euro Exchange Rates. 4th IAEE Asian Conference,
  Beijing, 2014.
}
}
