% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_ljung_box.R
\name{diag_ljung_box}
\alias{diag_ljung_box}
\title{Ljung-Box Test for Serial Correlation}
\usage{
diag_ljung_box(x, lags = c(8, 10, 12), order = 0, squared = FALSE)
}
\arguments{
\item{x}{A \code{matrix} / \code{data.frame} / \code{numeric vector} of
(multivariat) financial time series. Each column contains a series, each
row an observation of the series.}

\item{lags}{The number of lags of cross-correlation matrices used in the
tests. Can take multiple values. Defaults to \code{lags = c(8, 10, 12)}.}

\item{order}{If the test is performed on residuals of a fitted time series
model such as (V)ARMA or GARCH \code{order} generally equals the count of
coefficients from the fitted model. Defaults to \code{order = 0}.}

\item{squared}{Logical switch, if the squared series should be tested.
Defaults to \code{squared = FALSE}.}
}
\value{
The Ljung-Box Test statistic, it's \eqn{p}-value and further
  parameters as a \code{data.frame}.
}
\description{
The Ljung-Box Test for serial correlation.
}
\section{Details}{
 {

The univariate Ljung-Box (1978) Test is denoted by

\deqn{Q^{*}(m) = T(T + 2) \sum^{m}_{i=1} \hat{\rho}^{2}_{i} / (T - i)}{
Q*(m) = T * (T + 2) * \sum^m_i=1 * \rho²_i / (T - i)}

where \eqn{T} stands for the sample size and \eqn{\hat{\rho}_{i}}{\rho_i}
for the lag-\eqn{i} sample autocorrelation of \eqn{e_{t}}{e_t}.

The Hypothesis \eqn{H_{0} : \rho_{1} = ... = \rho_{m} = 0}{H0 :
\rho_1 = ... = \rho_m = 0} is tested against \eqn{H_{1} : \rho_{i} \neq
0}{H1 : \rho_i != 0} for \eqn{i = (1 \leq i \leq m)}{i = (1 \le i \le m)}.
Under the null hypothesis of no conditional heteroscedasticity in
\eqn{a_{t}}{a_t}, the test statistic \eqn{Q^{*}(m)}{Q*(m)} is
asymptotically distributed as \eqn{\chi^{2}_{m}}{\chi²_m}.

The generalization of the Ljung-Box Test to the multivariate case,
see e.g. Hosking (1980, 1981) is denoted by

\deqn{Q^{*}_{k}(m) = T^{2} \displaystyle\sum_{i=1}^{m} \frac{1}{T - i}
b_{i}^{\prime} (\hat{\rho}^{-1}_{0} \otimes \hat{\rho}^{-1}_{0})
b_{i}}{Q*_k(m) = T² * \sum_{i=1}^m * (1 / (T - i)) * b'_i(\rho_0^-1
\otimes \rho_0^-1) * b_i}

where \eqn{T} stands for the sample size, \eqn{k} for the dimension of
\eqn{a_{t}}{a_t} and \eqn{b_{i} = vec(\hat{\rho}^{\prime}_{i})}{b_i =
vec(\rho'_i)} with \eqn{\hat{\rho}_{j}}{\rho_j} being the
lag-\eqn{\hat{\rho}_{j}}{\rho_j} cross-correlation matrix of
\eqn{a^{2}_{t}}{a²_t}. As in the univariate case, under the null hypothesis
of no conditional heteroscedasticity in \eqn{a_{t}}{a_t},
\eqn{Q_{k}^{*}(m)}{Q*_k(m)} is asymptoically distributed as
\eqn{\chi^{2}_{k^{2}m}}{\chi²_{k²m}}.

}
}

\examples{
# create heteroscedastic data
dat <- mgarchBEKK::simulateBEKK(3, 500)
eps <- data.frame(eps1 = dat$eps[[1]], eps2 = dat$eps[[2]],
                  eps3 = dat$eps[[3]])

# perform multivariate arch test
diag_ljung_box(eps)

}
\references{
{

  Ljung, G. & Box, G. E. P  (1978). On a measure of lack of fit in time
  series models. Biometrika 66: 67-72.

  Hosking, J. R. M. (1980). The multivariate portmanteau statistic. Journal
  of the American Statistical Association, 75: 602–607.

  Hosking, J. R. M. (1981). Lagrange-multiplier tests of multivariate time
  series model. Journal of the Royal Statistical Society,
  Series B, 43: 219–230.

  Li, W. K. (2004). Diagnostic Checks in Time Series. Chapman & Hall / CRC.
  Boca Raton, FL.

  Mahdi, E. & McLeod, A. I. (2012). Improved multivariate portmanteau
  diagnostic test. Journal of Time Series Analysis 33/2, 211-222.

  Tsay, R. S. (2014). Multivariate Time Series Analysis with R
  and Financial Applications. John Wiley. Hoboken, NJ.

  Tsay, R. S. (2015). MTS: All-Purpose Toolkit for Analyzing
  Multivariate Time Series (MTS) and Estimating Multivariate Volatility
  Models. R package version 0.33.

}
}
\seealso{
\code{\link{diag_std_et}} for the transformation of a multivariate
  financial time series to a standardized scalar series which can be tested
  for conditional heteroscedasticity (ARCH effect) with the Ljung-Box Test,
  \code{\link{mv_ch_tests}} for different varieties of ARCH tests,
  \code{\link{diag_std_et_cnd}} for the transformation of a multivariate
  financial time series to a standardized scalar series and a multivariate
  (marginally) standardized series based on fitted conditional covariace
  matrices.
}
