% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_mv_ch_model.R
\name{diag_mv_ch_model}
\alias{diag_mv_ch_model}
\title{Multivariate Conditional Heteroscedasticity Model Checking}
\usage{
diag_mv_ch_model(x, lags = c(8, 10, 12), baq_err_cor = TRUE)
}
\arguments{
\item{x}{A \code{list} of two \code{data.frame}s / \code{matrices}
containing: \cr
 Element 1 - eps: Residuals for a k-dimensional asset return series
 (return series minus conditional mean / residuals of the mean-equation).\cr
 Element 2 - cnd_h: The fitted volatility matrices (estimated
 conditional covariance matrix). Has to have k^2 columns.\cr
 Can handle \code{baq_nif} class objects.}

\item{lags}{The number of lags used in the tests. Defaults to 10.}

\item{baq_err_cor}{Logical switch if potential negative variances should be
handled for \code{baq_nif} object. This is achieved by removing all
observations up to the last negative variance in descending order from eps
and cnd_h.}
}
\value{
Various test statistics and their p-values.
}
\description{
Apply four portmanteau test statistics to check the validity of a fitted
multivariate volatility model.
}
\details{
{

For the four test statistics employed check the details section of
\code{link{mv_ch_tests}}. For the transformation of \code{eps} & \code{cnd_h}
check the details section of \code{\link{diag_std_et_cnd}}. For an
comprehensive explanation see the references, especially Tsay (2014).

}
}
\examples{
# create heteroscedastic data
dat <- mgarchBEKK::simulateBEKK(2, 150)
eps <- data.frame(eps1 = dat$eps[[1]], eps2 =dat$eps[[2]])

# fit a GARCH model
gjr <- mGJR(eps[, 1], eps[, 2])

# conditional covariance matrices
cnd_h <- matrix(unlist(gjr$H.estimated), ncol = 4, byrow = TRUE)

#diagnostics on estimated conditional covariance matrices
diag_mv_ch_model(x = list(as.matrix(eps), cnd_h))

#alternative for a fitted baq_nif object
baq <- baq_nifunction(gjr)
diag_mv_ch_model(baq)

}
\references{
{

  Ljung G. & Box G. E. P. (1978). On a measure of lack of fit in time series
  models. Biometrika 66: 67-72.

  Dufour, J. M. & Roy R. (1985). The \eqn{t} copula and related copulas.
  Working Paper. Department of Mathematics, Federal Institute of Technology.

  Dufour, J. M. & Roy R. (1986). Generalized portmanteau statistics and tests
  of randomness. Communications in Statistics-Theory and Methods, 15:
  2953-2972.

  Ling, S. & Li, W. K. (1997). Diagnostic checking of nonlinear multivariate
  time series with multivariate ARCH errors. Journal of Time Series
  Analysis, 18: 447–464.

  Tse, Y. K. (2002). Residual-based diagnostics for conditional
  heteroscedasticity models. Econometric Journal, 5: 358–373.

  Tsay, R. S. (2014). Multivariate Time Series Analysis with R and Financial
  Applications. John Wiley. Hoboken, NJ.

  Tsay, R. S. (2015). MTS: All-Purpose Toolkit for Analyzing Multivariate
  Time Series (MTS) and Estimating Multivariate Volatility Models.
  R package version 0.33.

}
}
\seealso{
\code{\link{diag_std_et_cnd}}, \code{\link{diag_dufour_roy}},
  \code{\link{diag_ljung_box}}
}
