% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_std_et.R
\name{diag_std_et}
\alias{diag_std_et}
\title{Standardized Univariate Series \eqn{e_{t}}{e_t}}
\usage{
diag_std_et(x)
}
\arguments{
\item{x}{A \code{matrix} / \code{data.frame} / \code{numeric vector} of
(multivariat) financial time series. Each column contains a series, each
row an observation of the series.}
}
\value{
The standardized univariate series \eqn{e_{t}}{e_t} of \code{x} as
  a \code{matrix}.
}
\description{
Transform a \eqn{k}-dimensional series \eqn{a_{t}}{a_t} to
the standardized univariate series \eqn{e_{t}}{e_t}.
}
\section{Details}{
 {

A \eqn{k}-dimensional series \eqn{a_{t}}{a_t} can be transformed to a
standardized univariate series \eqn{e_{t}}{e_t}:

\deqn{e_{t} = a_{t}^{\prime} \textstyle{\sum}^{-1} a_{t} - k}{e_t = a'_t *
\sum^-1 * a_t - k}

where \eqn{\sum} denotes the unconditional covariance matrix of the
\eqn{k}-dimensional series \eqn{a_{t}}{a_t}.

}
}

\examples{
# create heteroscedastic data
dat <- mgarchBEKK::simulateBEKK(3, 500)
eps <- data.frame(eps1 = dat$eps[[1]], eps2 = dat$eps[[2]],
                  eps3 = dat$eps[[3]])

# transform to standardized univariate series e_t
et <- diag_std_et(eps)

}
\references{
{

  Dufour J. M. & Roy, R. (1985). The \eqn{t} copula and related copulas.
  Working Paper. Department of Mathematics, Federal Institute of Technology.

  Dufour J. M. & Roy, R. (1986). Generalized portmanteau statistics and tests
  of randomness. Communications in Statistics-Theory and Methods, 15:
  2953-2972.

  Tsay, R. S. (2014). Multivariate Time Series Analysis with R
  and Financial Applications. John Wiley. Hoboken, NJ.

  Tsay, R. S. (2015). MTS: All-Purpose Toolkit for Analyzing
  Multivariate Time Series (MTS) and Estimating Multivariate Volatility
  Models. R package version 0.33.

}
}
\seealso{
\code{\link{diag_dufour_roy}} & \code{\link{diag_ljung_box}} for Test
  Statistics for the transformed standardized series \code{diag_std_et}
}
