% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baq_nifunction.R
\name{is.baq_nif}
\alias{is.baq_nif}
\title{Reports whether x is a baq_nif object}
\usage{
is.baq_nif(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a baq_nif object
}
\keyword{internal}
