% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_std_et.R
\name{is.std_et}
\alias{is.std_et}
\title{Reports whether x is a std_et object}
\usage{
is.std_et(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a std_et object
}
\keyword{internal}
