% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_std_et_cnd.R
\name{is.std_et_cnd}
\alias{is.std_et_cnd}
\title{Reports whether x is a std_et_cnd object}
\usage{
is.std_et_cnd(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a std_et_cnd object
}
\keyword{internal}
