% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march_test.R
\name{march_test}
\alias{march_test}
\title{Multivariate ARCH test}
\usage{
march_test(zt, m = 10)
}
\arguments{
\item{zt}{nT-by-k data matrix of a k-dimensional financial time series,
each column contains a series.}

\item{m}{The number of lags of cross-correlation matrices used in the
tests.}
}
\value{
Various test statistics and their p-values.
}
\description{
Performs tests to check the conditional heteroscedasticity in a vector
time series.
}
\examples{
# create heteroscedastic data
dat <- mgarchBEKK::simulateBEKK(2, 500)
eps <- data.frame(eps1 = dat$eps[[1]], eps2 =dat$eps[[2]])

# perform multivariate arch test
march_test(eps)

}
\references{
{
  Tsay (2014, Chapter 7). Multivariate Time Series Analysis with R and
  Financial Applications. John Wiley. Hoboken, NJ.

  Ruey S. Tsay (2015, Page 26-28). MTS: All-Purpose Toolkit for Analyzing
  Multivariate Time Series (MTS) and Estimating Multivariate Volatility
  Models. R package version 0.33.

}
}
