% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mch_diag.R
\name{mch_diag}
\alias{mch_diag}
\title{Multivariate Conditional Heteroscedasticity Model Checking}
\usage{
mch_diag(x, m = 10, baq_err_cor = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{A \code{list} of two \code{data.frame}s / \code{matrices}
containting: \cr
 Element 1 - eps: A T-by-k matrix of residuals for a k-dimensional asset
 return series (return series minus conditional mean / residuals of the mean-
 equation). \cr
 Element 2 - est_ccovm: The fitted volatility matrices (estimated
 conditional covariance matrix). The dimension is a T-by-k^2 matrix. \cr
 Handles \code{baq_nif} class objects on its own.}

\item{m}{The number of lags used in the tests. Defaults to 10.}

\item{baq_err_cor}{Logical switch if potential negative variances should be
handled for \code{baq_nif} object. This is achieved by removing all
observations up to the last negative variance in descending order from eps
and ccovm.}

\item{quiet}{Logical switch if information about applying the mch_diag
function should be shown in the console.}
}
\value{
Various test statistics and their p-values.
}
\description{
Apply four portmanteau test statistics to check the validity of a fitted
multivariate volatility model.
}
\details{
This function is a copy of \code{\link[MTS]{MCHdiag}}. It deviates
  from the originial in two ways:

 1) it displays the result as a \code{data.frame} and

 2) thus makes it possible to save the results to a variable.

 The test-statistics are from Tsays(2014) Multivariate Time Series as follows:

 Q(m):
}
\examples{
# create heteroscedastic data
dat <- mgarchBEKK::simulateBEKK(2, 500)
eps <- data.frame(eps1 = dat$eps[[1]], eps2 =dat$eps[[2]])

# fit a GARCH model
gjr <- mGJR(eps[, 1], eps[, 2])

# conditional covariance matrices
ccovm <- as.data.frame(matrix(unlist(gjr$H.estimated), ncol = 4, byrow = T))

#diagnostics on estimated conditional covariance matrices
mch_diag(x = list(eps, ccovm))

#alternative for a fitted baq_nif object
baq <- baq_nifunction(gjr)
mch_diag(gjr)

}
\references{
{
  Ruey S. Tsay (2014, Chapter 7). Multivariate Time Series Analysis with R and
  Financial Applications. John Wiley. Hoboken, NJ.

  Ruey S. Tsay (2015, Page 26-28). MTS: All-Purpose Toolkit for Analyzing
  Multivariate Time Series (MTS) and Estimating Multivariate Volatility
  Models. R package version 0.33.

}
}
