% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baq_nifunction.R
\name{par_list_to_dataframe}
\alias{par_list_to_dataframe}
\title{Transforms a list of matrices to a data.frame.}
\usage{
par_list_to_dataframe(x, colname_ext = "", colname_ext_sep = " ")
}
\arguments{
\item{x}{A list of matrices to be transformed (est.params or
asy.se.coeff from a calss mGJR object)}

\item{colname_ext}{Add an extension to the parameter names.
Defaults to nothing ("").}

\item{colname_ext_sep}{Seperator for colname extension. Will be
ignored if no extension was added.}
}
\description{
Transforms a list of matrices to a data.frame.
}
\keyword{internal}
