% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baq_nifunction.R
\name{print.baq_nif}
\alias{print.baq_nif}
\title{Reports whether x is a baq_nif object}
\usage{
\method{print}{baq_nif}(x, ...)
}
\arguments{
\item{x}{An object to \code{\link[base]{print}}}
}
\description{
Reports whether x is a baq_nif object
}
\keyword{internal}
