% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baq_nifunction.R
\name{baq_nifunction}
\alias{baq_nifunction}
\title{Apply a News Impact Function}
\usage{
baq_nifunction(mgjrclass, epsnames = c("eps1", "eps2"), er_grid = 10,
  er_grid_by = 0.2)
}
\arguments{
\item{mgjrclass}{A mGJR class object.}

\item{epsnames}{Custom names for the variables eps1 and eps2. Character
vector of length two. Defaults to c("eps1", "eps2").}

\item{er_grid}{The minimum and maximum value for the hypothetical returns
in the News Impact Function. Numerical vector of length one.}

\item{er_grid_by}{Steps by which to construct the sequence of returns.
Numerical vector of length one.}
}
\value{
The applied News Impact Function packaged as a \code{baq_nif} class
  object. The News Impact Functions' conditional variance/correlation in
  long- (\code{data.frame}) and and wide-format (\code{list} of
  \code{matrices}).
 The values are defined as:
\describe{
  \item{eps}{\code{data.frame} with eps1 and eps2 inherited from the
  \code{mGJR} class object.}
  \item{est.params}{\code{list} of \code{matrices} with the baqGARCH
    parameters needed for the News Impact Function.}
  \item{baqH_long}{\code{data.frame} containing the News Impact on the
  conditional variance/correlation of eps1/eps2 in long-format.}
  \item{baqH_wide}{A \code{list} of \code{matrices} with the News Impact on
   the conditional variance of eps1/eps2 and conditional correlation in
   wide-format.}
}
}
\description{
Apply a News Impact Function to a mGJR class object.
}
\examples{
{
# create data
eps <- mgarchBEKK::simulateBEKK(2, 100)

# fit the model
gjr <- mgarchBEKK::mGJR(eps$eps[[1]], eps$eps[[2]])

# apply the news impact function to the model
nif <- baq_nifunction(gjr)
}
}
\references{
{
  H. Schmidbauer & A. Roesch. Volatility Spillovers Between Crude Oil
  Prices. International Conference on Policy Modeling. EcoMod,
  Berlin, 2008.

  H. Schmidbauer & A. Roesch. Volatility Spillovers Between Crude Oil
  Prices and Us Dollar To Euro Exchange Rates. 4th IAEE Asian Conference,
  Beijing, 2014.
}
}
