% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baq_nifunction.R
\name{baq_nifunction}
\alias{baq_nifunction}
\title{News Impact on Conditional Variance}
\usage{
baq_nifunction(x, epsnames = c("series1", "series2"), er_grid = 10,
  er_grid_by = 0.2, ni_long = TRUE, ni_wide = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{A mGJR class object.}

\item{epsnames}{Custom names for the variables eps1 and eps2. Character
vector of length two. Defaults to c("eps1", "eps2").}

\item{er_grid}{The minimum and maximum value for the hypothetical returns
in the news impact Function. Numerical vector of length one.}

\item{er_grid_by}{Steps by which to construct the sequence of returns.
Numerical vector of length one.}

\item{ni_long}{Logical switch if the conditional variance after news impact
should be returned in long format (as \code{data.frame}).}

\item{ni_wide}{Logical switch if the conditional variance after news impact
should be returned in wide format (as \code{matrix}).}

\item{quiet}{Logical switch if information about applying baq_nifunction
should be shown in the console.}
}
\value{
The applied news impact Function packaged as a \code{baq_nif} class
  object. The news impact functions' conditional variance / correlation
  (optional) as long- (\code{data.frame}) and and wide-format (\code{list}
  of \code{matrices}).
 The values are defined as:
\describe{
  \item{series_names}{character \code{vector} with the names for eps1 and
    eps2}
  \item{eps}{\code{data.frame} with eps1 and eps2 inherited from the
    \code{mGJR} class object.}
  \item{baq_h}{\code{data.frame} with the estimated conditional covariance
    matrices inherited from the \code{mGJR} class object. Each column stands
    for a matrix element of the conditional covariance matrix, each row for
    the time at which it was observed. The time notation of the rows is
    identical to \code{eps}.}
  \item{coef}{\code{data.frame} with the baqGARCH coefficients
    needed for the news impact Function.}
  \item{coef_se}{\code{data.frame} with the baqGARCH coefficients
    standard errors.}
  \item{coef_tval}{\code{data.frame} with the baqGARCH coefficients
    T-values.}
  \item{baq_ni_cndh_long}{\code{data.frame} containing the news impact on
    the conditional variance / correlation of eps1 / eps2 in long-format.}
  \item{baq_ni_cndh_wide}{A \code{list} of \code{matrices} with the news
    impact on the conditional variance of eps1/eps2 and conditional
    correlation in wide-format.}
}
}
\description{
Apply a news impact Function to a mGJR class object.
}
\section{Details}{
{

Schmidbauer & Roesch (2008, 2014) proposed a bivariate asymmetric quadratic
GARCH (baqGARCH) model to account for asymmetric components in bivariate
volatility matrices. They also apply a news impact on the conditional volatility
of a fitted baqGARCH model by letting the conditional volatility matrices
\eqn{H = (h_{ij})} depend on \eqn{x = (x_{1}, x_{2})}{x = (x_1, x_2)}:

\deqn{x \mapsto H(x) = C^{\prime}C + A^{\prime}xx^{\prime}A +
B^{\prime}\sum B + S_{w}(x) * \Gamma^{\prime}xx^{\prime}\Gamma}{x -> H(x) =
C'C + A'xx'A + B'\sumB + S_w(x) * \Gamma'xx'\Gamma}

where \eqn{\sum} is the unconditional covariance matrices of the bivariate
time series and \eqn{x} the vector of potential innovations (i.e. returns) in
the bivariate series affecting it's  conditional volatility.

The contour lines (conditional variance after news impact) are based on the
functions:

\deqn{x \mapsto h_{11}(x), x \mapsto h_{22}(x), x \mapsto h_{12}(x)/
\sqrt{h_{11}(x)h_{22}(x)}}{x -> h_{11}(x), x -> h_{22}(x), x -> h_{12}
(x)/\sqrt(h_{11}(x) * h_{22}(x)),}

where the function \eqn{x_{11}} stands for the news impact on the next day's
conditional variance of returns on series 1, \eqn{x_{22}} stands for the news
impact on the next day's conditional variance of returns on series 2 and
\eqn{h_{12}(x)/\sqrt{h_{11}(x) * h_{22}(x)}} for the conditional correlation
of returns (series 1 & 2).
}
}

\examples{
# create data
eps <- mgarchBEKK::simulateBEKK(2, 100)

# fit the model
gjr <- mgarchBEKK::mGJR(eps$eps[[1]], eps$eps[[2]])

# apply the news impact function to the model
nif <- baq_nifunction(gjr)
}
\references{
{
  Schmidbauer, H. & Roesch, A. (2008). Volatility Spillovers Between Crude
  Oil Prices. International Conference on Policy Modeling. EcoMod, Berlin.

  Schmidbauer, H. & Roesch, A. (2014). Volatility Spillovers Between Crude
  Oil Prices and Us Dollar To Euro Exchange Rates. 4th IAEE Asian
  Conference,Beijing.
}
}
\seealso{
\code{\link{baq_niplot}} to plot the news impact function and
 \code{\link{diag_mv_ch_model}} to perform tests to determine model adequacy
}
