% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baq_niplot.R
\name{baq_niplot}
\alias{baq_niplot}
\title{Plot News Impact on Conditional Variance}
\usage{
baq_niplot(x, ni_plots = c("var1", "var2", "cor"), plotdir = "horizontal")
}
\arguments{
\item{x}{A baq_nif class object.}

\item{ni_plots}{A character vector specifying which conditional correlation
/ variance to plot. Can contain any combination of "var1", "var2" and
"cor". Defaults to c("var1", "var2", "cor").}

\item{plotdir}{A character vector of length one to determine if the plots
should be created in a row ("horizontal") or a column ("vertical").
Defaults to "horizontal". Currently the output doesn't look very nice,
since \code{par()} is not set up for vertical yet.}
}
\value{
A plot of the news impact on the conditional variance / correlation.
}
\description{
Plot the News Impact on the conditional variance / correlation of a baq_nif
class object.
}
\section{Details}{
{

For more details on the baqGARCH model and the application of a news impact
function as proposed by Schmidbauer & Roesch (2008, 2014) check the details
section of \code{\link{baq_nifunction}} and further see the references.

}
}

\examples{
# create data
eps <- mgarchBEKK::simulateBEKK(2, 100)

# fit the model
gjr <- mgarchBEKK::mGJR(eps$eps[[1]], eps$eps[[2]])

# apply the news impact function to the model
nif <- baq_nifunction(gjr)

# plot the results
baq_niplot(nif)

# plot only the the first series
baq_niplot(nif, ni_plots = "var1")

\dontrun{

# save plot as pdf
pdf("baqplot_example.pdf", width = 15, height = 7.2)
baq_niplot(nif)
dev.off()
}
}
\references{
{
  Schmidbauer, H. & Roesch, A. (2008). Volatility Spillovers Between Crude
  Oil Prices. International Conference on Policy Modeling. EcoMod, Berlin.

  Schmidbauer, H. & Roesch, A. (2014). Volatility Spillovers Between Crude
  Oil Prices and Us Dollar To Euro Exchange Rates. 4th IAEE Asian
  Conference, Beijing.

  Nychka, D. & Furrer, R. & Paige, J. & Sain, S. (2017). “fields: Tools for
  spatial data.” doi: 10.5065/D6W957CT (URL:http://doi.org/10.5065/D6W957CT),
  R package version 9.6.
}
}
\seealso{
\code{\link{baq_nifunction}} for fitting a \code{baq_nif}-class
  object
}
